requirements.txt
environs==14.5.0
python-telegram-bot==13.15
redis==3.2.1
requests==2.32.5
